requirements/base.txt
django
django-allauth
psycopg2
django-bootstrap4